Extract functions to init_session module
